Adds new method run_in_device() to GoclKernel
This is the asynchronous version.

The asynchronous pattern used is very experimental. The method returns a
GoclEvent which can be used to synchronize with other OpenCL queued calls
that accept an event wait list.

To execute code when event triggers, GoclEvent::then() method should be
used.

This pattern is somehow similar to the Promise pattern usually found in
Javascript libraries.